// src/enums/page-path.enum.ts
export enum PagePath {
  LOGIN = "/Login",
  REGISTER = "/Register",
  VERIFY_EMAIL = "/VerifyEmail",
  VERIFY_OTP = "/VerifyOTP",
  RESET_PASSWORD = "/ResetPassword",
  NOT_FOUND = "/404",
  HOME = "/Home",
  WORK_VOLUME = "/Home/Workvolume",
  USER_MANAGEMENT = "/Home/User",
  USER_DETAIL = "/Home/User/:id",
  SKIN_TYPE_MANAGEMENT = "/Home/SkintypeManagement",
  SERVICE_MANAGEMENT = "/Home/ServiceManagement",
  SLOT_MANAGEMENT = "/Home/SlotManagement",
  HOME_PAGE = "/Homepage",
  SHOPPING_PAGE = "/Homepage/Shoppingpage",
  QUIZ = "/Homepage/QuizTest",
  SKIN_SERVICE = "/Homepage/Service",
  // SKIN_SERVICE_DETAIL = "/Homepage/Service/:serviceId",
  SKIN_SERVICE_DETAIL = "/Homepage/Service/Detail",
  BLOG = "/Homepage/Blog",
  // BLOG_DETAIL = "/Homepage/Blog/:id",
  BLOG_DETAIL = "/Homepage/Blog/Detail",
  SKIN_THERAPIST = "/Homepage/SkinTherapist",
  SKIN_THERAPIST_DETAIL = "/Homepage/SkinTherapist/Detail",
  PRICE_SERVICE = "/Homepage/Price",
  BOOKING_SERVICE = "/Homepage/BookingService",
  SKIN_TYPE = "/Homepage/SkinType",
  COMPLETE_RESULT = "/Homepage/CompleteBooking",
  BOOKING_INFO_CONFIRM = "/Homepage/BookingInfoConfirm",
  CUSTOMER_PROFILE = "/Homepage/Profile",
  CUSTOMER_BOOKING_DETAIL = "/Homepage/CustomerBookingDetail",
  ANY = "*",
  FORBIDDEN = "/403",
  PROFILE = "/Home/Profile",
  BOOKING = "/Home/Booking",
  // BOOKING_DETAIL = "/Home/Booking/:bookingId",
  BOOKING_DETAIL = "/Home/Booking/Detail",
  QUIZ_MANAGEMENT = "/Home/QuizManagement",
  BLOG_MANAGEMENT = "/Home/BlogManagement",
  SCHEDULE_FOR_STAFF_MANAGEMENT = "/Home/Schedule",
  SCHEDULE_FOR_THERAPIST = "/Home/TherepistSchedule",
  CREATE_BLOG = "/Home/CreateBlog",
  THERAPIST_MANAGEMENT = "/Home/TherapistManagement",
  ROOT = "/",
  // TICKET_MANAGEMENT = "/ticket-management",
  // TIMEKEEPING_DETAIL = "/timekeeping/:code/detail",
  // TIMEKEEPING_MANAGEMENT = "/timekeeping",
  // USER_DETAIL = "/user-management/:id",
  // USER_MANAGEMENT = "/user-management",
}